fix: resolve the four AppleClang 21 -Wimplicit-int-float-conversion sites
The macos-latest runner move brought AppleClang 21, which diagnoses four
pre-existing int-widened-inside-a-float-expression sites. Each now carries the
explicit (float) cast that spells out the conversion the compiler was already
performing implicitly:

  src/PluginEditor.cpp:245   roundToInt ((float) inner.getWidth() * 0.40f)
  src/PluginEditor.cpp:246   roundToInt ((float) getWidth()       * 0.40f)
  src/gui/LookAndFeel.cpp:262  x0 + (float) k * (barW + gap)
  src/dsp/VelvetNoise.cpp:30   std::round ((float) m * cell + ...)

Nothing is suppressed: no #pragma, no -Wno-..., no change to
juce_recommended_warning_flags. The cast is also the surrounding idiom -- the
two lines above the VelvetNoise site already read (float) decorrSamps /
(float) maxTaps.

Behaviour is verified unchanged rather than argued. int * float already
converts the int operand to float under the usual arithmetic conversions, so
(float) i * f is the same expression written out; each of the three translation
units was compiled at the shipped flags with -g0 -fno-lto appended (so debug
metadata and LTO bitcode cannot mask the comparison) before and after the edit,
and all three objects are BYTE-IDENTICAL. That covers the whole translation
unit, a stronger statement than the scenario-matrix twin dump -- which matters
for VelvetNoise.cpp, where it makes the velvet tap grid bit-exact (Class A).

Reproduced and re-checked locally under Clang 18 at the shipped flags: the same
four warnings fire at the same line:column as AppleClang 21 does on CI, and all
four are gone after the change with no other warning added or removed.

juce::Rectangle::proportionOfWidth(0.40f) was considered for the PluginEditor
sites and rejected: it returns ValueType (w * p), a truncation where the
existing code rounds.

Docs: CI_CD's toolchain paragraph recorded these as unfixed and now records the
fix and its evidence; the KNOWN_ISSUES and FUTURE_RISKS v0.9.4 version-sync
headers pointed at it for the same four diagnostics; DOCUMENTATION_COVERAGE
gains the audit entry. No CHANGELOG entry (CHANGELOG_POLICY rule 3 -- nothing a
user of the plug-in can observe changed) and no ARCHITECTURE_REVIEW_GATE item.

+++ src/PluginEditor.cpp
@@ -242,8 +242,8 @@ void AnamorphAudioProcessorEditor::ABControl::paint (juce::Graphics& g)
     auto inner = getLocalBounds();
     g.setFont (juce::Font (juce::FontOptions (14.0f)).withExtraKerningFactor (0.04f));
 
-    auto aRect = inner.removeFromLeft (juce::roundToInt (inner.getWidth() * 0.40f));
-    auto bRect = inner.removeFromRight (juce::roundToInt (getWidth() * 0.40f));
+    auto aRect = inner.removeFromLeft (juce::roundToInt ((float) inner.getWidth() * 0.40f));
+    auto bRect = inner.removeFromRight (juce::roundToInt ((float) getWidth() * 0.40f));
 
     g.setColour (colours::textDim.withAlpha (0.7f));
     g.drawText ("/", inner, juce::Justification::centred);
